VLM Image Examples (#1064)
## Purpose ##
* Provide sample generation with images in VLM examples

## Prerequisites ##
* #1065

## Changes ##
* Replace sample generation text prompt with one that also contains an
image of a cat, similar to the example provided by
[llava-1.5](https://huggingface.co/nm-testing/llava-1.5-7b-hf-FP8-dynamic/discussions/1)
# <img width="50%" alt="tool icon"
src="http://images.cocodataset.org/train2017/000000231895.jpg" />

## Testing ##
* The sample generation was run without oneshot in order to save time

Llava
```
========== SAMPLE GENERATION ==============
USER:
Please describe the animal in this image
 ASSISTANT: A white kitten is laying on a computer keyboard.
==========================================
```

Mllama
```
========== SAMPLE GENERATION ==============
user

Please describe the animal in this image
assistant

The image features a small white cat curled up on a computer keyboard, which is situated on a desk in front of a computer monitor. The cat's fur is short and white, and it is lying on its side, with its head resting on the keyboard and its body curled up behind it. The keyboard is white and has a standard QWERTY layout, with the cat positioned on the right-hand side.

In the background, a computer monitor is visible, along with various office supplies and kn
==========================================
```

Pixtral
```
========== SAMPLE GENERATION ==============
Setting `pad_token_id` to `eos_token_id`:2 for open-end generation.
Please describe the animal in this image
The image features a white cat comfortably resting on a white computer keyboard. The cat appears relaxed and is lying down with its head resting on the keys. The keyboard is placed on a wooden desk. In the background, there are various items on the desk, including a computer monitor, a yellow toy, a small bowl, and a red apple. The overall scene suggests a cozy and homely environment where the cat feels at ease.
==========================================
```

---------

Signed-off-by: Kyle Sayers <kylesayrs@gmail.com>
Co-authored-by: Dipika Sikka <dipikasikka1@gmail.com>
Signed-off-by: Kyle Sayers <kylesayrs@gmail.com>