Require 'ready' label for transformers tests (#1079)
SUMMARY:
This PR moves the `transformers-tests` job to its own workflow which
only runs on a PR that has the 'ready' label. With this implementation,
if the PR already has the 'ready' label when a new commit is pushed, the
new commit will re-run the job. I also updated the PR comment workflow
to note that the 'ready' label must be added to run this job.

There are some visual warts/caveats that come with these triggers
because the filtering cannot be done by the triggers, only the jobs:

Whenever a label is added or a commit is pushed, the workflow will still
be triggered. However, if the PR does not have the 'ready' label, the
`transformers-tests` job will be skipped. While the job being skipped
means the runner/resources will not actually be used, it also means that
the workflow/job will still show up in various places in the GitHub UI
(Actions tab, PR checks). It will show as skipped, but it will still
allow the green check mark because that appears as long as there are no
failures.

Basically, this means that reviews/people able to merge will need to pay
closer attention to the jobs run on the PR to make sure that the
`transformers-tests` job was run/not skipped.

TEST PLAN:
Should affect this PR.

---------

Signed-off-by: Domenic Barbuzzi <domenic@neuralmagic.com>