Turn off 2:4 sparse compression until supported in vllm (#1092)
This PR temporarily disables the newly added Sparse24 compression
feature in example script, as support for this feature is not yet
available in vLLM.

Support for Sparse24 compression is being added in vLLM via [this
PR](https://github.com/vllm-project/vllm/pull/12097). Once that PR is
merged, this change will be reverted to re-enable the feature.

Signed-off-by: Rahul Tuli <rahul@neuralmagic.com>