[Bug Fix] Fix test that requre GPU (#1096)
SUMMARY:
Nightly tests uses gpus, ci tests dont. 
In the set up of
`tests/llmcompressor/transformers/obcq/test_consecutive_runs.py`, a bug
was found where `quantization_config` was always passed. This will
result in an error if the model from the config file is not optimized.
We are passing in dense model, so it fails.


TEST PLAN:
Run the failing test, make sure it passes!