Replace readme paths with urls (#1097)
## Purpose ##
* Files with the `.md` extension are not listed in the
[MANIFEST.in](https://github.com/vllm-project/llm-compressor/blob/main/MANIFEST.in),
meaning that they will not be included in the LLM Compressor pypi
package. This means that references to these files are left dangling for
users who have installed from the pypi package. Rather than including
`.md` in the package and having to also ship all the large images files
associated with them, this PR moves the references to urls hosted by
github
* While the github url paths may change between versions, this solution
works in lieu of a dedicated readthedoc build for each version
* This solution also aligns with the practice of other libraries which
point to hosted urls rather than file paths
* Note that this does not apply to files which are themselves `.md`
files, as these files will not be included in the pypi distribution
  * `src/llmcompressor/transformers/finetune/README.md`
  * `src/llmcompressor/pipelines/sequential/README.md`

## Changes ##
* Replace readme file paths with urls
* Small change to `DisableQuantization` to better catch cases where
exceptions such as tracing exceptions are triggered

## Testing ##
* N/A

---------

Signed-off-by: Kyle Sayers <kylesayrs@gmail.com>