Fix: Re-enable Sparse Compression for 2of4 Examples (#1153)
This PR restores sparse compression for our `2of4` examples, which was
previously disabled due to a bug in the vLLM Cutlass integration.

#### Background  
A bug in the Cutlass integration caused certain sparse-only compressed
models to produce gibberish results. To mitigate this issue, we
temporarily turned off sparse compression for our `2of4` examples.

The bug has since been fixed by @tlrmchlsmth in
[vllm-project/vllm#13198](https://github.com/vllm-project/vllm/pull/13198).
With this fix in place, we can safely re-enable sparse compression for
these examples.

#### Changes  
- Re-enable sparse compression for `2of4` examples.  

#### Testing  
- Verified that sparse-only compressed models now produce expected
outputs.

---------

Signed-off-by: Rahul Tuli <rahul@neuralmagic.com>
Co-authored-by: Dipika Sikka <dipikasikka1@gmail.com>