[Bugfix] Staged 2of4 example (#1238)
## Purpose ##
* Fix staged 2of4 example

## Background ##
* When #1160 landed, this change introduced a bug in the recipe
container which meant that the recipe was not recompiled after
`append`ing. This caused sgpt to initialize twice and gptq to never
initialize, leading to a sparsity-only quantization config
* At some point, a changed was introduced which causes previous stages
to become reconstructed after recipe recompilation. This means that
without resetting the session in between stages, previous stages will
initialize twice.
* In order to avoid this issue, this PR introduces `session.reset()` in
between stages
* This change has the consequence of creating `recipe.yaml` files which
do not have the full recipe history. However, I believe this is
acceptable for the time being, as the stage runner and this work flow
will be removed in the next release.

---------

Signed-off-by: Kyle Sayers <kylesayrs@gmail.com>