[Tests] Fix oneshot + finetune test by passing splits to oneshot (#1316)
## Purpose ##
* Fix test which was failing because the calibration split was not
explicitly passed to the oneshot call. While normally this would not be
a problem, as `make_dataset_splits` will auto-infer which split is the
calibration split, the `make_dataset_splits` function does its inference
using the split name `train`, whereas ultrachat uses the spilt name
`train_sft`.

## Changes ##
* Pass `splits` argument to oneshot and train calls. This also reduces
the runtime of the test significantly.

---------

Signed-off-by: Kyle Sayers <kylesayrs@gmail.com>