replace custom version handling with setuptools-scm
Use setuptools-scm to automatically compute a version based on the
latest git tag.

Remove the old logic for reading a static file to get the version
string, but configure setuptools-scm to generate a similar file for
anyone who wants to import it to get the version information.

Replace the build_type value set in code with an environment variable
so setup.py can continue to change the versioning scheme based on the
type of build being performed.

When the BUILD_TYPE is "nightly", produce an alpha release for the
_next_ version, with the date as the alpha version. For example, when
the most recent tag is 0.4.1 the nightly release version built on 3
Apr 2025 is 0.4.2a20250403.

When the BUILD_TYPE is "release", if the commit being built has a tag
use that tag's value as the version.

When the BUILD_TYPE is "dev", or if the commit being built does not
have a tag, or the git repository has pending changes, compute a
development version combining the distance since the last tag with the
expected next tag. For example, if the most recent tag is 0.4.1 and
there have been 74 commits since that tag and the most recent commit
SHA is 5e2c1c92, then the version would be 0.4.2.dev74+g5e2c1c92. If
the local git repository has modified files that are not committed,
the date is appended in the local version field, giving
0.4.2.dev74+g5e2c1c92.d20250403.

When the BUILD_TYPE is "release", depend on the pinned version of
compressed-tensors. Otherwise, let the version float so that dev and
CI environments can control it.

Signed-off-by: Doug Hellmann <dhellmann@redhat.com>