[Compression] Update sparsity calculation lifecycle when fetching the compressor (#1332)
# Summary
- Updating the lifecycle based on feedback from research and meta
- #1293 is addressed @eldarkurtic 

### 1. No longer always infer sparsity 
- If there is existing sparsity in the model that should be accounted
for while compressing the model, the user is required to indicate this
by providing a sparsity config. This is not a common case and primarily
something we've seen from research. As a result, it does not need to be
the default pathway
- If no sparsity config is provided, only check for sparsity if a recipe
with sparsity was applied OR the user explicitly would like to apply a
sparsity compressor by setting `skip_compression_stats` parameter
- Based on these changes, we no longer use the global_sparsity to
dictate whether or not a sparse compressor is applied

### 2. Change the name of `skip_compression_stats` to
`skip_sparsity_compression_stats` and set to True by default
### 3. Rename `sparsity_config` to `sparsity_metadata_config` for
clarity
### 4. Add additional log outputs to make it a lot more clear as to what
steps are running
### 5. Fix log outputs to prevent cutting off

# Follow-ups:
- Need to update passing saving arguments properly - they are currently
not being passed apart from save_compressed and there is no clear API on
what the arguments are/how they should be passed, both to
save_pretrained and when providing an output dir to oneshot

# Testing
- Tested a sparse-only model
- Tested a model with both sparsity and quantization
- Tested a model with existing sparsity
- Tested a model with only quantization

For all cases, the model compresses to the expected format.

---------

Co-authored-by: Brian Dellabetta <brian-dellabetta@users.noreply.github.com>