[Tracing] Remove `TraceableWhisperForConditionalGeneration` (#1310)
## Purpose ##
* Remove unneeded traceable definition. I believe that I mistook
WhisperForConditionalGeneration as needing a traceable definition during
testing. The offending line attempts to check the shape of the
`input_features`, however the shape of `input_features` is known via
metadata, so this is a valid and non-conditional check.
* This may be a result of updated attention definitions which allow
shapes to be traced through the model with more clarity

```python3
if input_features.shape[-1] != expected_seq_length
    raise ValueError
```

## Changes ##
* Remove `TraceableWhisperForConditionalGeneration`

## Testing ##
* Ran `examples/multimodal_audio/whisper_example.py` to completion and
verified model outputs

Signed-off-by: Kyle Sayers <kylesayrs@gmail.com>