AWQ resolved mappings -- ensure shapes align (#1372)
SUMMARY:
Add a robustness check to AWQ to exclude mappings where the layer shapes
don't align. This is a known issue, just wanted to do it in a different
PR because it occurs in a different location in AutoAWQ, and I wanted to
keep the initial AWQ PR to be as close to the AutoAWQ implementation so
the git history reflects our changes well.


TEST PLAN:
Added unit test to check default case and edge cases

---------

Signed-off-by: Brian Dellabetta <bdellabe@redhat.com>
Co-authored-by: Dipika Sikka <dipikasikka1@gmail.com>