LM Eval tests -- ignore vision tower for VL fp8 test (#1562)
SUMMARY:
The current lm-eval test for vision language models with fp8_dynamic
scheme include the vision tower component of the model in the
compression. As discussed with @anmarques and @eldarkurtic, this is
generally not a good idea and we want to err on the side of accuracy
over improved runtime when the tradeoff exists. Excluding vision tower
from compression slightly decreases accuracy (from 0.866 to 0.833) in
this case, but generally speaking it will degrade performance and we
don't want to encourage users to do so.

This PR updates the test to explicitly ignore the vision tower
components in the test


TEST PLAN:
Ran test a few times locally, reproduced 0.8333 value each time.
Confirmed the model size is now slightly larger with the vision tower
excluded from compression

---------

Signed-off-by: Brian Dellabetta <bdellabe@redhat.com>