Enable interactive debugging in llm-compressor by skipping self-references in locals (#1617)
If one tries to run interactive debugging session in e.g. VSCode,
LLM-Compressor would crash with the following error:

```
ValueError: Some keys are not used by the HfArgumentParser: ['kwargs', 'local_args']
```

This happens because dictionary `local_args = locals()` contains
`local_args` and `kwargs` as keys (due to different execution context in
debugpy compared to the standard "script mode" with `python
quantize.py`), which HF argument parser in OneShot can't handle. This
doesn't happens when LLM-Compressor is used in a "script mode" with
`python quantize_my_model.py`.

This patch makes it explicit that `local_args` and `kwargs` shouldn't be
in local_args and makes interactive session with LLM-Compressor
possible.

---------

Co-authored-by: Brian Dellabetta <brian-dellabetta@users.noreply.github.com>