[Examples] Remote `trust_remote_code` from people's speech dataset (#1654)
## Purpose ##
* Remove `trust_remote_code` option from peoples_speech dataset loading
in whisper example
  * Squelch `datasets` warning

```
`trust_remote_code` is not supported anymore.
Please check that the Hugging Face dataset 'MLCommons/peoples_speech' isn't based on a loading script and remove `trust_remote_code`.
If the dataset is based on a loading script, please ask the dataset author to remove it and convert it to a standard format like Parquet.
```

## Background ##
As of
[datasets==4.0.0](https://github.com/huggingface/datasets/releases/tag/4.0.0),
datasets with remote code are no longer supported

## Changes ##
* Remove `trust_remote_code_data` from data loading in whisper example

Signed-off-by: Kyle Sayers <kylesayrs@gmail.com>
Co-authored-by: Dipika Sikka <dipikasikka1@gmail.com>