[AWQ] fixes to matching logic and #1742 bugfix (#1759)
SUMMARY:
- [x] I introduced a bug in #1742 that caused `lm_eval` awq test to
fail. This reverts that change, re-setting the original state dict of
the parent module in the grid search for best scales.
- [x] This also updates to the new module matching API, excluding from
the resolved mappings any modules that match the list in `ignore`. This
should resolve a user issue with command-a-vision, which has k_proj etc.
layers in the vision_encoder that we want to exclude in our resolved
mappings.


TEST PLAN:
awq lm_eval test is passing now. Running command-a-vision check

---------

Signed-off-by: Brian Dellabetta <bdellabe@redhat.com>