Remove redundant compat-pytorch-1_9-pytorch-tests (#1798)
SUMMARY:
Currently, we have a test titled `compat-pytorch-1_9-pytorch-tests`
which runs the exact same code as the test above `pytorch-tests`. The
only difference is the python version used (3.10 vs 3.11). Based on the
test name it seems like it was previously used for testing
`pytorch==1.9` support (which no longer applies).

This pr removes the redundant test.

Signed-off-by: Fynn Schmitt-Ulms <fschmitt@redhat.com>