[Entrypoints] initialize processor error handling (#1796)
SUMMARY:
Resolves #1795
* #1795 

Currently, we initialize a processor in entrypoint `pre_process` even if
one isn't provided, even though it isn't needed for data-free recipes
like `FP8_DYNAMIC` or `W4A16`, causing downstream user issues like
#1795. This updates pre-processing to
- wrap processor initialization in a try/catch
- error out if initialization fails and a processor is required (i.e. if
a dataset is needed for training/calibration)
- otherwise, log a warning if an output_dir is provided, because the
processor will not be saved with the trained/compressed model.


TEST PLAN:
Example script in #1795 succeeds on this branch, confirmed error is
raised if `output_dir` is set and error is raised if `dataset` is set.

---------

Signed-off-by: Brian Dellabetta <bdellabe@redhat.com>