Use correct model (#1810)
SUMMARY:
This test was skipped and when run, was giving the can't copy out of
meta tensor error. This is because the
[model](https://huggingface.co/Xenova/llama2.c-stories110M/discussions/4)
is missing lm_head and embed_token layers and therefore only loaded to
meta device. Switching to our model to avoid this issue.


TEST PLAN:
Tested locally and passed.

Signed-off-by: shanjiaz <zsjwpianpian@gmail.com>
Co-authored-by: Brian Dellabetta <brian-dellabetta@users.noreply.github.com>