Remove `isort`, `black` and `flake8`, replace with `ruff` (#1808)
SUMMARY:
Ruff supports import sorting, formatting, and linting. We already use if
for formatting (`black` is listed as a dev dependency but isn't used)
and linting.

In this pr:
- 6af05b191ba171a0cc66c92088a768a2d5a8226b
- Cleans up `pyproject.toml`, removing old configs/updating ruff configs
to also check import sorting
- Updates `Makefile` to just call `ruff` in `make quality` and `make
style`
  - Removes `black`, `flake8`, and `isort` from dev dependencies
  - Updates dev docs description of `make quality` command
- 796a5b8632e88ca0d33e14737fba684ef97aa863
- Replaces `flake8: noqa` file wide markers with `ruff: noqa` (not
strictly necessary since `ruff` will respect the `flake8: noqa` markers
but for cleanliness)
  - Removes unnecessary `isort: skip_file` markers

Note: this is builds off changes from #1807

---------

Signed-off-by: Fynn Schmitt-Ulms <fschmitt@redhat.com>