[bugfix] use raw string for regexes when backslash is used (#1823)
SUMMARY:
Resolves #1776 

The `gemma3_example.py` script was not correctly escaping the `\`
character in its regexes. This adds the `r"` prefix to the string, so
that python knows it is a raw string that needs to be escaped, avoiding
warnings.

It appears to be isolated to this example, I couldn't find other regexes
with `\` characters that weren't being properly escaped.


TEST PLAN:
Example runs successfully after changes.

Signed-off-by: Brian Dellabetta <bdellabe@redhat.com>