Add concurrency marker to transformers workflow (#1824)
SUMMARY:
Currently if you open a pr and mark it as ready the transformers check
workflow will run. If you then push more commits to the pr, it will
re-run the workflow but won't cancel the existing workflow run.

Since we only care about the results of the most recent push to the pr
branch, this change will make it so that we cancel pre-existing runs
that have the same branch.

Ref:
https://docs.github.com/en/actions/how-tos/write-workflows/choose-when-workflows-run/control-workflow-concurrency#example-using-concurrency-and-the-default-behavior


Note: only adding this to the transformers workflow because it takes the
longest (~30 mins) and uses custom runners.

Signed-off-by: Fynn Schmitt-Ulms <fschmitt@redhat.com>
Co-authored-by: Brian Dellabetta <brian-dellabetta@users.noreply.github.com>