[Performance] Speed up DataLoader with Multiple DataLoader Workers (#1821)
when the dataset is too large, it should be speeded up to min(8x,cpu//2)
times

SUMMARY:

When the dataset is large, using the default DataLoader is too slow. In
my experiment, it took 1.5 hours to prepare the cache. After setting
num_workers to min(8, cpu // 2), the time was reduced to 7 minutes.

TEST PLAN:

Tested with a 5K dataset, comparing the performance between this patch
and the original implementation.

---------

Signed-off-by: Baoyuan Qi <qibaoyuan@126.com>
Co-authored-by: Brian Dellabetta <brian-dellabetta@users.noreply.github.com>