update examples to avoid zstandard dependency (#1839)
SUMMARY:
Previously, the AWQ examples used the
[pile-val](https://huggingface.co/datasets/mit-han-lab/pile-val-backup)
dataset, a vestige from the original paper. Pile-val is saved in a .zst
format and requires `zstandard`, which is not a required dependency of
`llm-compressor`. This PR updates the example to use the huggingface
ultrachat dataset, which is used in our other examples and has no
`zstandard` requirement.


TEST PLAN:
Examples ran locally. output of llama model: 🔥 😄 

> <|begin_of_text|>Hello my name is Sarah and I'm a 27-year-old writer
from the United States. I've always been fascinated by the world of
technology and the impact it has on our daily lives. I'm excited to be
here and share my thoughts and insights with you.
In my free time, I enjoy reading science fiction and fantasy novels, as
well as watching movies and TV shows. I'm also a bit of a gamer and
enjoy playing strategy games like Civilization and Starcraft.

Signed-off-by: Brian Dellabetta <bdellabe@redhat.com>